Fix letfovers handlers autentification by roles

+++ PGP.Application/Comments/Commands/DeleteComment/DeleteCommentCommandHandler.cs
@@ -1,8 +1,8 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using PGP.Application.Common.Interfaces;
 using PGP.Application.Exceptions;
 using PGP.Persistence;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -11,10 +11,12 @@ namespace PGP.Application.Comments.Commands.DeleteComment
     public class DeleteCommentCommandHandler : IRequestHandler<DeleteCommentCommand, Unit>
     {
         private readonly IPGPDbContext _context;
+        private readonly ICurrentUserService _currentUserService;
 
-        public DeleteCommentCommandHandler(IPGPDbContext context)
+        public DeleteCommentCommandHandler(IPGPDbContext context, ICurrentUserService currentUserService)
         {
             _context = context;
+            _currentUserService = currentUserService;
         }
 
         public async Task<Unit> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
@@ -25,6 +27,11 @@ namespace PGP.Application.Comments.Commands.DeleteComment
             if (comment == null)
             {
                 throw new NotFoundException($"Comment id {request.Id} not exists.");
+            }
+
+            if (_currentUserService.UserId != comment.CreatedByUserId && !_currentUserService.Role.Equals("Admin"))
+            {
+                throw new UnauthorizedException("You can't do this action.");
             }
 
             _context.Comments.Remove(comment);
